+++ src/lib/ebus/data.cpp
@@ -50,35 +50,6 @@ static const char* defaultTemplateFieldMap[] = {
 };
 
 
-
-string getDataFieldName(const string& name, bool* supportsLanguage) {
-  if (name.find("name") != string::npos || name.find("field") != string::npos) {
-    return "name";
-  }
-  if (name.find("part") != string::npos) {
-    return "part";
-  }
-  if (name.find("type") != string::npos) {
-    return "type";
-  }
-  if (name.find("divisor") != string::npos) {
-    if (name.find("values") != string::npos) {
-      *supportsLanguage = true;
-      return "divisor/values";
-    }
-    return "divisor";
-  }
-  *supportsLanguage = true;
-  if (name == "values" || name == "unit") {
-    return name;
-  }
-  if (name.find("comment") != string::npos) {
-    return "comment";
-  }
-  return "";
-}
-
-
 string AttributedItem::formatInt(size_t value) {
   ostringstream stream;
   stream << dec << static_cast<unsigned>(value);
@@ -107,8 +78,8 @@ void AttributedItem::dumpString(bool prependFieldSeparator, const string& str, o
 }
 
 void AttributedItem::appendJson(bool prependFieldSeparator, const string& name, const string& value,
-    bool asString, ostream* output) {
-  bool plain = !asString;
+    bool forceString, ostream* output) {
+  bool plain = !forceString && !value.empty();
   if (plain) {
     plain = value == "false" || value == "true";
     if (!plain) {
@@ -125,7 +96,15 @@ void AttributedItem::appendJson(bool prependFieldSeparator, const string& name, 
   if (plain) {
     *output << value;
   } else {
-    *output << '"' << value << '"';
+    *output << '"';
+    if (value.find_first_of('"') == string::npos) {  // check for nested '"'
+      *output << value;
+    } else {
+      string replaced = value;
+      replace(replaced.begin(), replaced.end(), '"', '\'');
+      *output << replaced;
+    }
+    *output << '"';
   }
 }
 
@@ -138,8 +117,13 @@ void AttributedItem::mergeAttributes(map<string, string>* attributes) const {
   }
 }
 
-void AttributedItem::dumpAttribute(bool prependFieldSeparator, const string& name, ostream* output) const {
-  dumpString(prependFieldSeparator, getAttribute(name), output);
+void AttributedItem::dumpAttribute(bool prependFieldSeparator, bool asJson, const string& name, ostream* output)
+    const {
+  if (asJson) {
+    appendJson(prependFieldSeparator, name, getAttribute(name), false, output);
+  } else {
+    dumpString(prependFieldSeparator, getAttribute(name), output);
+  }
 }
 
 bool AttributedItem::appendAttribute(OutputFormat outputFormat, const string& name, bool onlyIfNonEmpty,
@@ -169,10 +153,19 @@ bool AttributedItem::appendAttributes(OutputFormat outputFormat, ostream* output
     for (const auto entry : m_attributes) {
       ret = true;
       if (!entry.second.empty() && entry.first != "unit" && entry.first != "comment") {
+        const string& key = entry.first;
         if (outputFormat & OF_JSON) {
-          appendJson(true, entry.first, entry.second, false, output);
+          if (key == "zz" || key == "qq") {
+            result_t result = RESULT_EMPTY;
+            size_t addr = parseInt(entry.second.c_str(), 16, 0, 255, &result);
+            if (result == RESULT_OK) {
+              *output << FIELD_SEPARATOR << " \"" << key << "\": " << addr;
+              continue;
+            }
+          }
+          appendJson(true, key, entry.second, false, output);
         } else {
-          *output << " " << entry.first << "=" << entry.second;
+          *output << " " << key << "=" << entry.second;
         }
       }
     }
@@ -464,27 +457,44 @@ result_t SingleDataField::create(const string& name, const map<string, string>& 
   return RESULT_OK;
 }
 
-void SingleDataField::dumpPrefix(ostream* output) const {
+void SingleDataField::dumpPrefix(bool prependFieldSeparator, bool asJson, ostream* output) const {
   *output << setw(0) << dec;  // initialize formatting
-  dumpString(false, m_name, output);
-  *output << FIELD_SEPARATOR;
-  if (m_partType == pt_masterData) {
-    *output << "m";
-  } else if (m_partType == pt_slaveData) {
-    *output << "s";
+  if (asJson) {
+    if (prependFieldSeparator) {
+      *output << FIELD_SEPARATOR;
+    }
+    *output << "\n     {";
+    appendJson(false, "name", m_name, true, output);
+  } else {
+    dumpString(prependFieldSeparator, m_name, output);
   }
   *output << FIELD_SEPARATOR;
+  if (asJson) {
+    *output << " \"slave\": " << (m_partType == pt_slaveData ? "true" : "false") << ", ";
+  } else {
+    if (m_partType == pt_masterData) {
+      *output << "m";
+    } else if (m_partType == pt_slaveData) {
+      *output << "s";
+    }
+  }
+  if (!asJson) {
+    *output << FIELD_SEPARATOR;
+  }
+  m_dataType->dump(asJson, m_length, true, output);
 }
 
-void SingleDataField::dumpSuffix(ostream* output) const {
-  dumpAttribute(true, "unit", output);
-  dumpAttribute(true, "comment", output);
+void SingleDataField::dumpSuffix(bool asJson, ostream* output) const {
+  dumpAttribute(true, asJson, "unit", output);
+  dumpAttribute(true, asJson, "comment", output);
+  if (asJson) {
+    *output << "}";
+  }
 }
 
-void SingleDataField::dump(ostream* output) const {
-  dumpPrefix(output);
-  m_dataType->dump(m_length, true, output);
-  dumpSuffix(output);
+void SingleDataField::dump(bool prependFieldSeparator, bool asJson, ostream* output) const {
+  dumpPrefix(prependFieldSeparator, asJson, output);
+  dumpSuffix(asJson, output);
 }
 
 result_t SingleDataField::read(const SymbolString& data, size_t offset,
@@ -499,7 +509,7 @@ result_t SingleDataField::read(const SymbolString& data, size_t offset,
   if (offset + (remainder?1:m_length) > data.getDataSize()) {
     return RESULT_ERR_INVALID_POS;
   }
-  if (isIgnored() || (fieldName != NULL && (m_name != fieldName || fieldIndex > 0))) {
+  if (isIgnored() || (fieldName != NULL && m_name != fieldName) || fieldIndex > 0) {
     return RESULT_EMPTY;
   }
   result_t res = m_dataType->readRawValue(offset, m_length, data, output);
@@ -519,7 +529,7 @@ result_t SingleDataField::read(const SymbolString& data, size_t offset,
   if (offset + (remainder?1:m_length) > data.getDataSize()) {
     return RESULT_ERR_INVALID_POS;
   }
-  if (isIgnored() || (fieldName != NULL && (m_name != fieldName || fieldIndex > 0))) {
+  if (isIgnored() || (fieldName != NULL && m_name != fieldName) || fieldIndex > 0) {
     return RESULT_EMPTY;
   }
   bool shortFormat = outputFormat & OF_SHORT;
@@ -527,16 +537,20 @@ result_t SingleDataField::read(const SymbolString& data, size_t offset,
     if (leadingSeparator) {
       *output << ",";
     }
-    if (!shortFormat) {
-      *output << "\n    ";
+    if (fieldIndex < 0 && !shortFormat) {
+      *output << "\n     ";
     }
     if (outputIndex >= 0 || m_name.empty() || !(outputFormat & OF_NAMES)) {
-      *output << "\"" << static_cast<signed int>(outputIndex < 0 ? 0 : outputIndex) << "\":";
+      if (fieldIndex < 0) {
+        *output << "\"" << static_cast<signed int>(outputIndex < 0 ? 0 : outputIndex) << "\":";
+      }
       if (!shortFormat) {
         *output << " {\"name\": \"" << m_name << "\"" << ", \"value\": ";
       }
     } else {
-      *output << "\"" << m_name << "\":";
+      if (fieldIndex < 0) {
+        *output << "\"" << m_name << "\":";
+      }
       if (!shortFormat) {
         *output << " {\"value\": ";
       }
@@ -681,10 +695,19 @@ result_t ValueListDataField::derive(const string& name, PartType partType, int d
   return RESULT_OK;
 }
 
-void ValueListDataField::dump(ostream* output) const {
-  dumpPrefix(output);
-  if (!m_dataType->dump(m_length, true, output)) {  // no divisor appended
-    bool first = true;
+void ValueListDataField::dump(bool prependFieldSeparator, bool asJson, ostream* output) const {
+  dumpPrefix(prependFieldSeparator, asJson, output);
+  // no divisor appended since it is not allowed for ValueListDataField
+  bool first = true;
+  if (asJson) {
+    *output << ", \"values\": {";
+    for (const auto it : m_values) {
+      appendJson(!first, formatInt(it.first), it.second, true, output);  // TODO optimize?
+      first = false;
+    }
+    *output << " }";
+  } else {
+    *output << FIELD_SEPARATOR;
     for (const auto it : m_values) {
       if (first) {
         first = false;
@@ -693,8 +716,8 @@ void ValueListDataField::dump(ostream* output) const {
       }
       *output << it.first << "=" << it.second;
     }
-  }  // else: impossible since divisor is not allowed for ValueListDataField
-  dumpSuffix(output);
+  }
+  dumpSuffix(asJson, output);
 }
 
 result_t ValueListDataField::readSymbols(const SymbolString& input, size_t offset,
@@ -788,12 +811,16 @@ result_t ConstantDataField::derive(const string& name, PartType partType, int di
   return RESULT_OK;
 }
 
-void ConstantDataField::dump(ostream* output) const {
-  dumpPrefix(output);
-  if (!m_dataType->dump(m_length, true, output)) {  // no divisor appended
-    *output << (m_verify?"==":"=") << m_value;
-  }  // else: impossible since divisor is not allowed for ConstantDataField
-  dumpSuffix(output);
+void ConstantDataField::dump(bool prependFieldSeparator, bool asJson, ostream* output) const {
+  dumpPrefix(prependFieldSeparator, asJson, output);
+  // no divisor appended since it is not allowed for ConstantDataField
+  if (asJson) {
+    appendJson(false, "value", m_value, true, output);
+    *output << ", \"verify\":" << (m_verify ? "true" : "false");
+  } else {
+    *output << FIELD_SEPARATOR << (m_verify?"==":"=") << m_value;
+  }
+  dumpSuffix(asJson, output);
 }
 
 result_t ConstantDataField::readSymbols(const SymbolString& input, size_t offset,
@@ -905,14 +932,26 @@ size_t DataFieldSet::getLength(PartType partType, size_t maxLength) const {
 }
 
 string DataFieldSet::getName(ssize_t fieldIndex) const {
-  if (fieldIndex < 0) {
+  if (fieldIndex < (ssize_t)m_ignoredCount) {
     return m_name;
   }
-  if ((size_t)fieldIndex >= m_fields.size()) {
+  if ((size_t)fieldIndex + m_ignoredCount >= m_fields.size()) {
     return "";
   }
   if (m_uniqueNames) {
-    return m_fields[fieldIndex]->getName(-1);
+    if (m_ignoredCount == 0) {
+      return m_fields[fieldIndex]->getName(-1);
+    }
+    ssize_t remain = fieldIndex;
+    for (const auto field : m_fields) {
+      if (field->isIgnored()) {
+        continue;
+      }
+      remain--;
+      if (remain == 0) {
+        return field->getName(-1);
+      }
+    }
   }
   ostringstream ostream;
   ostream << static_cast<signed>(fieldIndex);
@@ -946,15 +985,10 @@ bool DataFieldSet::hasField(const char* fieldName, bool numeric) const {
   return false;
 }
 
-void DataFieldSet::dump(ostream* output) const {
-  bool first = true;
+void DataFieldSet::dump(bool prependFieldSeparator, bool asJson, ostream* output) const {
   for (const auto field : m_fields) {
-    if (first) {
-      first = false;
-    } else {
-      *output << FIELD_SEPARATOR;
-    }
-    field->dump(output);
+    field->dump(prependFieldSeparator, asJson, output);
+    prependFieldSeparator = true;
   }
 }
 
@@ -978,7 +1012,7 @@ result_t DataFieldSet::read(const SymbolString& data, size_t offset,
     if (result != RESULT_EMPTY) {
       found = true;
     }
-    if (findFieldIndex && fieldName == field->getName(-1)) {
+    if (findFieldIndex && !field->isIgnored() && (fieldName == NULL || fieldName == field->getName(-1))) {
       if (fieldIndex == 0) {
         if (!found) {
           return RESULT_ERR_NOTFOUND;
@@ -999,7 +1033,7 @@ result_t DataFieldSet::read(const SymbolString& data, size_t offset,
 result_t DataFieldSet::read(const SymbolString& data, size_t offset,
     bool leadingSeparator, const char* fieldName, ssize_t fieldIndex,
     OutputFormat outputFormat, ssize_t outputIndex, ostream* output) const {
-  bool previousFullByteOffset = true, found = false, findFieldIndex = fieldName != NULL && fieldIndex >= 0;
+  bool previousFullByteOffset = true, found = false, findFieldIndex = fieldIndex >= 0;
   if (outputIndex < 0 && (!m_uniqueNames || ((outputFormat & OF_JSON) && !(outputFormat & OF_NAMES)))) {
     outputIndex = 0;
   }
@@ -1025,7 +1059,7 @@ result_t DataFieldSet::read(const SymbolString& data, size_t offset,
       found = true;
       leadingSeparator = true;
     }
-    if (findFieldIndex && fieldName == field->getName(-1)) {
+    if (findFieldIndex && !field->isIgnored() && (fieldName == NULL || fieldName == field->getName(-1))) {
       if (fieldIndex == 0) {
         if (!found) {
           return RESULT_ERR_NOTFOUND;