+++ src/lib/ebus/datatype.cpp
@@ -46,17 +46,27 @@ using std::setw;
 using std::endl;
 
 
-bool DataType::dump(size_t length, bool appendSeparatorDivisor, ostream* output) const {
-  *output << m_id;
-  if (isAdjustableLength()) {
-    if (length == REMAIN_LEN) {
-      *output << ":*";
+bool DataType::dump(bool asJson, size_t length, bool appendDivisor, ostream* output) const {
+  if (asJson) {
+    *output << "\"type\": \"" << m_id << "\"" << FIELD_SEPARATOR << " \"isbits\": "
+            << (getBitCount() < 8 ? "true" : "false") << FIELD_SEPARATOR << " \"length\": ";
+    if (isAdjustableLength() && length == REMAIN_LEN) {
+      *output << "-1";
     } else {
-      *output << ":" << static_cast<unsigned>(length);
+      *output << static_cast<unsigned>(length);
     }
-  }
-  if (appendSeparatorDivisor) {
-    *output << FIELD_SEPARATOR;
+  } else {
+    *output << m_id;
+    if (isAdjustableLength()) {
+      if (length == REMAIN_LEN) {
+        *output << ":*";
+      } else {
+        *output << ":" << static_cast<unsigned>(length);
+      }
+    }
+    if (appendDivisor) {
+      *output << FIELD_SEPARATOR;
+    }
   }
   return false;
 }
@@ -506,21 +516,27 @@ size_t NumberDataType::calcPrecision(int divisor) {
   return precision;
 }
 
-bool NumberDataType::dump(size_t length, bool appendSeparatorDivisor, ostream* output) const {
+bool NumberDataType::dump(bool asJson, size_t length, bool appendDivisor, ostream* output) const {
   if (m_bitCount < 8) {
-    DataType::dump(m_bitCount, appendSeparatorDivisor, output);
+    DataType::dump(asJson, m_bitCount, appendDivisor, output);
   } else {
-    DataType::dump(length, appendSeparatorDivisor, output);
+    DataType::dump(asJson, length, appendDivisor, output);
   }
-  if (!appendSeparatorDivisor) {
+  if (!appendDivisor) {
     return false;
   }
   if (m_baseType) {
     if (m_baseType->m_divisor != m_divisor) {
+      if (asJson) {
+        *output << ", \"divisor\": ";
+      }
       *output << (m_divisor / m_baseType->m_divisor);
       return true;
     }
   } else if (m_divisor != 1) {
+    if (asJson) {
+      *output << ", \"divisor\": ";
+    }
     *output << m_divisor;
     return true;
   }
@@ -856,7 +872,8 @@ result_t NumberDataType::writeSymbols(size_t offset, size_t length, istringstrea
         dvalue = round(dvalue * m_divisor);
       }
       if (hasFlag(SIG)) {
-        if (dvalue < -(1LL << (8 * length)) || dvalue >= (1LL << (8 * length))) {
+        if (dvalue < -exp2((8 * static_cast<double>(length)) - 1)
+            || dvalue >= exp2((8 * static_cast<double>(length)) - 1)) {
           return RESULT_ERR_OUT_OF_RANGE;  // value out of range
         }
         if (dvalue < 0 && m_bitCount != 32) {
@@ -865,7 +882,7 @@ result_t NumberDataType::writeSymbols(size_t offset, size_t length, istringstrea
           value = static_cast<int>(dvalue);
         }
       } else {
-        if (dvalue < 0.0 || dvalue >= (1LL << (8 * length))) {
+        if (dvalue < 0.0 || dvalue >= exp2(8 * static_cast<double>(length))) {
           return RESULT_ERR_OUT_OF_RANGE;  // value out of range
         }
         value = (unsigned int)dvalue;