+++ src/lib/ebus/test/test_message.cpp
@@ -87,10 +87,10 @@ int main() {
     {"r,cir,Status01,VL/RL/AussenTemp/VLWW/SpeicherTemp/Status,,08,B511,01,,,temp1;temp1;temp2;temp1;temp1;pumpstate", "28.0;24.0;4.938;35.0;41.0;4", "ff08b5110101", "093830f00446520400ff", "d"},
     {"r,message circuit,message name,message comment,,25,B509,0d2800,,,tempsensor", "temp=-14.00 Temperatursensor [Temperatur];sensor=ok [Fühlerstatus]", "ff25b509030d2800", "0320ff00", "D"},
     {"r,message circuit,message name,message comment,,25,B509,0d2800,,,tempsensor,,field unit,field comment", "temp=-14.00 field unit [field comment];sensor=ok [Fühlerstatus]", "ff25b509030d2800", "0320ff00", "D"},
-    {"r,message circuit,message name,message comment,,25,B509,0d2800,,,tempsensor,,field unit,field comment", "\n    \"temp\": {\"value\": -14.00},\n    \"sensor\": {\"value\": \"ok\"}", "ff25b509030d2800", "0320ff00", "j"},
-    {"r,message circuit,message name,message comment,,25,B509,0d2800,,,tempsensor,,field unit,field comment", "\n    \"temp\": {\"value\": -14.00, \"unit\": \"field unit\", \"comment\": \"field comment\"},\n" "    \"sensor\": {\"value\": \"ok\", \"comment\": \"Fühlerstatus\"}", "ff25b509030d2800", "0320ff00", "J"},
+    {"r,message circuit,message name,message comment,,25,B509,0d2800,,,tempsensor,,field unit,field comment", "\n     \"temp\": {\"value\": -14.00},\n     \"sensor\": {\"value\": \"ok\"}", "ff25b509030d2800", "0320ff00", "j"},
+    {"r,message circuit,message name,message comment,,25,B509,0d2800,,,tempsensor,,field unit,field comment", "\n     \"temp\": {\"value\": -14.00, \"unit\": \"field unit\", \"comment\": \"field comment\"},\n" "     \"sensor\": {\"value\": \"ok\", \"comment\": \"Fühlerstatus\"}", "ff25b509030d2800", "0320ff00", "J"},
     {"r,message circuit,message name,message comment,,25,B509,0d2800,,,temp,,field unit,field comment,,,sensor", "temp=-14.00 field unit [field comment];sensor=ok [Fühlerstatus]", "ff25b509030d2800", "0320ff00", "D"},
-    {"r,message circuit,message name,message comment,,25,B509,0d2800,,,D2C,,°C,Temperatur,,,sensor", "\n    \"0\": {\"name\": \"\", \"value\": -14.00},\n    \"1\": {\"name\": \"sensor\", \"value\": \"ok\"}", "ff25b509030d2800", "0320ff00", "j"},
+    {"r,message circuit,message name,message comment,,25,B509,0d2800,,,D2C,,°C,Temperatur,,,sensor", "\n     \"0\": {\"name\": \"\", \"value\": -14.00},\n     \"1\": {\"name\": \"sensor\", \"value\": \"ok\"}", "ff25b509030d2800", "0320ff00", "j"},
     {"r,cir,name,,,25,B509,0d2800,,,tempsensorc", "-14.00", "ff25b509030d2800", "0320ff55", ""},
     {"r,cir,name,,,25,B509,0d28,,m,sensorc,,,,,,temp", "-14.00", "ff25b509030d2855", "0220ff", ""},
     {"u,cir,first,,,fe,0700,,x,,bda", "26.10.2014", "fffe07000426100614", "00", "p"},
@@ -145,6 +145,25 @@ int main() {
     {"r,,x,,,,,6a00,,,UCH,10,bar,,Bit6,,BI6:1,0=B60;1=B61,,,Bit7,,BI7:1,0=B70;1=B71", "1.9;B60;B70", "ff08b509030d6a00", "02133f", "d" },
     {"*r,cir*cuit#level,na*me,com*ment,ff,75,b509,0d", "", "", "", "" },
     {"r,CIRCUIT,NAME,COMMENT,,,,0100,field,,UCH", "r,cirCIRCUITcuit,naNAMEme,comCOMMENTment,ff,75,b509,0d0100,field,s,UCH,,,: field=42", "ff75b509030d0100", "012a", "DN"},
+    {"r,CIRCUIT,NAME,COMMENT,,,,0100,field,,UCH",
+        // "\"naNAMEme\": {r,cirCIRCUITcuit,naNAMEme,comCOMMENTment,ff,75,b509,0d0100,field,s,UCH,,,: field=42"
+      "\n"
+      "   \"naNAMEme\": {\n"
+      "    \"name\": \"naNAMEme\",\n"
+      "    \"passive\": false,\n"
+      "    \"write\": false,\n"
+      "    \"lastup\": *,\n"
+      "    \"qq\": 255,\n"
+      "    \"zz\": 117,\n"
+      "    \"id\": [181, 9, 13, 1, 0],\n"
+      "    \"fields\": {\n"
+      "     \"0\": {\"name\": \"field\", \"value\": 42}\n"
+      "    },\n"
+      "    \"fielddefs\": [\n"
+      "     { \"name\": \"field\", \"slave\": true, \"type\": \"UCH\", \"isbits\": false, \"length\": 1, \"unit\": \"\", \"comment\": \"\"}\n"
+      "    ]\n"
+      "   }: \n"
+      "     \"field\": {\"value\": 42}", "ff75b509030d0100", "012a", "jN"},
   };
   templates = new DataFieldTemplates();
   unsigned int lineNo = 0;
@@ -349,8 +368,23 @@ int main() {
         message->storeLastData(*mstrs[index], *sstrs[index]);
       }
       ostringstream output;
-      if (withMessageDump && !decodeJson) {
-        message->dump(NULL, true, &output);
+      if (withMessageDump) {
+        if (decodeJson) {
+          message->decodeJson(false, false, false, OF_JSON|OF_DEFINTION, &output);
+          string str = output.str();
+          size_t start = str.find("\"lastup\": ");
+          if (start != string::npos) {
+            start += 10;
+            size_t end = str.find(",", start);
+            if (end != string::npos) {
+              str = str.substr(0, start)+"*"+str.substr(end);
+            }
+          }
+          output.str("");
+          output << str;
+        } else {
+          message->dump(NULL, true, &output);
+        }
         output << ": ";
       }
       result = message->decodeLastData(false, NULL, -1,
@@ -362,6 +396,7 @@ int main() {
         continue;
       }
       cout << "  \"" << check[2] << "\" / \"" << check[3] <<  "\": decode OK" << endl;
+      string outStr = output.str();
       bool match = inputStr == output.str();
       verify(false, "decode", check[2] + "/" + check[3], match, inputStr, output.str());
       if (checkUpdateTime || checkSameChangeTime) {